# src/RestApp/Controllers/CharacterController.cs
/*
 * Star Wars API
 *
 * This is a REST API for managing Star Wars characters
 *
 * OpenAPI spec version: 1.0.0
 * Contact: luk@mysl.tech
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using RestApp.Attributes;
using RestApp.Data.Contracts;
using RestApp.Models;
using Swashbuckle.AspNetCore.Annotations;

namespace RestApp.Controllers
{
    /// <summary>
    /// 
    /// </summary>
    public class CharacterController : BaseController
    {
        private readonly ICharacterService characterService;

        /// <summary>
        /// CharacterController
        /// </summary>
        /// <param name="characterService">reference to instance of CharacterRepository</param>
        public CharacterController(ICharacterService characterService)
        {
            this.characterService = characterService;
        }

        /// <summary>
        /// adds an character item
        /// </summary>
        /// <remarks>Adds an item character to the system</remarks>
        /// <param name="characterModel">Character item to add</param>
        /// <response code="201">item created</response>
        /// <response code="400">invalid input, object invalid</response>
        /// <response code="409">an existing item already exists</response>
        [HttpPost]
        [Route("/api_v1/characters")]
        [ValidateModelState]
        [SwaggerOperation("AddCharacter")]
        public virtual IActionResult AddCharacter([FromBody]CharacterModelDataContract characterModel)
        {
            var result = this.characterService.Create(characterModel);
            if (result == null) return this.NoContent();
            if (result.ResultIsOk==false)
            {
                return this.BadRequest(result.Error);
            }
            return new OkObjectResult(result);
        }

        /// <summary>
        /// Deletes a character
        /// </summary>
        
        /// <param name="characterId">Character id to delete</param>
        /// <response code="400">Invalid ID supplied</response>
        /// <response code="404">Character not found</response>
        [HttpDelete]
        [Route("/api_v1/characters/{characterId}")]
        [ValidateModelState]
        [SwaggerOperation("DeleteCharacter")]
        public virtual IActionResult DeleteCharacter([FromRoute][Required]int characterId)
        {
            var result = this.characterService.Delete(characterId);
            if (result == null) return this.NotFound();
            if (result.ResultIsOk == false)
            {
                return this.BadRequest(result.Error);
            }
            return new OkObjectResult(result);
        }

        /// <summary>
        /// get list of characters
        /// </summary>
        /// <remarks>Returns all characters of Star Wars movie </remarks>
        /// <param name="searchString">pass an optional search string for looking up characters</param>
        /// <param name="skip">number of records to skip for pagination</param>
        /// <param name="limit">maximum number of records to return</param>
        /// <response code="200">search results matching criteria</response>
        /// <response code="404">Characters not found</response>
        /// <response code="400">bad input parameter</response>
        [HttpGet]
        [Route("/api_v1/characters")]
        [ValidateModelState]
        [SwaggerOperation("GetListOfCharacters")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<CharacterModelDataContract>), description: "search results matching criteria")]
        public virtual IActionResult GetListOfCharacters([FromQuery]string searchString, [FromQuery]int? skip, [FromQuery][Range(0, 50)]int? limit)
        {
            var result = this.characterService.Read(searchString,skip,limit);
            if (result == null) return this.NotFound();
            return new OkObjectResult(result);
        }

        /// <summary>
        /// Update an existing character
        /// </summary>
        /// <param name="body">Character object that needs to be added first</param>
        /// <param name="characterId">Character id to update</param>
        /// <response code="400">Invalid ID supplied</response>
        /// <response code="404">Character not found</response>
        /// <response code="405">Validation exception</response>
        [HttpPut]
        [Route("/api_v1/characters/{characterId}")]
        [ValidateModelState]
        [SwaggerOperation("UpdateCharacter")]
        public virtual IActionResult UpdateCharacter([FromBody]CharacterModelDataContract body, [FromRoute][Required]int characterId)
        {
            if (characterId != body.Id) return this.ValidationProblem();
            else
            {
                var result = this.characterService.Update(body);
                if (result == null) return this.NotFound();
                if (result.ResultIsOk == false)
                {
                    return this.BadRequest(result.Error);
                }

                return new OkObjectResult(result);
            }
        }
    }
}